Move tab command parser helpers from ParserUtil to TabCommandParser
